Fix tool configuration object shape mismatch
Not sure where the best place to fix this is, but here is the problem:
The `addToolForElement` function takes a `configuration` argument which is then merged with the default config here: https://github.com/cornerstonejs/cornerstoneTools/blob/8e63394b0656774c86292507c2910f2b23b0abb5/src/tools/ZoomTool.js#L36

The problem is that it looks like most tools `defaultConfig` object has a sub-property `.configuration` that is what should be set by the `addToolForElement` call. So in it's current state, the configured properties don't actually have any effect because they are added to the top level of the `defaultConfig` object instead of the sub-property. This change fixes it in the passed parameter, but not sure if it's better to change at the config merge level for each tool? Not sure what the original intent is.